chore(app): update react ecosystem
