+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,23 @@
+---
+name: Feature request
+about: Suggest an idea for the nOS Client
+
+---
+
+<!-- A clear and concise description of feature you'd like to see and why. -->
+## Describe the solution you'd like and the expected behavior
+
+
+
+<!-- A clear and concise description of what the problem is.  -->
+## Is your feature request related to a problem? Please describe.
+
+
+
+<!--- If you have an idea or reference of how to implement this, suggest it here -->
+## Possible implementation / References
+
+
+
+<!-- Add any other context or screenshots about the feature request here. -->
+## Additional context
