+++ .npmignore
@@ -1,2 +1,3 @@
 CODE_OF_CONDUCT.md
 .github/
+.circleci/
