+++ es6/bindings.js
@@ -1,4 +1,4 @@
-import { useFallback, executor, exists, nos } from "./lib";
+import { executor } from "./lib";
 
 // Wallet Actions
