+++ es6/react/inject.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+import nos from "../";
+
+const injectNOS = Component => props => (
+  <Component nos={nos} {...props} />
+);
+
+export { injectNOS };
