+++ lib/index.js
@@ -3,18 +3,14 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.react = exports.default = void 0;
+exports.default = void 0;
 
 var _bindings = require("./bindings");
 
 var _lib = require("./lib");
 
-var reactNamespace = _interopRequireWildcard(require("./react"));
-
-function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj.default = obj; return newObj; } }
-
 /**
- * @nos/api-functions
+ * @nosplatform/api-functions
  * API bindings (with fallback) and types for nOS dApps
  * @author Jeroen Peeters <contact@nos.io>
  */
@@ -29,5 +25,3 @@ var _default = {
   testInvoke: _bindings.testInvoke
 };
 exports.default = _default;
-var react = reactNamespace;
-exports.react = react;
