+++ lib/lib.js
@@ -19,7 +19,7 @@ var executor = function executor(func, fallback, config) {
   return exists ? nos[func](config) : useFallback(fallback, config);
 };
 /**
- * @function useFallback 
+ * @function useFallback
  * @param {function} fallback checks if fallback function exists, otherwise return nothing
  * @param {object} args potential arguments that are passed along
  */
