+++ lib/react/index.js
@@ -3,17 +3,21 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.default = void 0;
+Object.defineProperty(exports, "nosProps", {
+  enumerable: true,
+  get: function get() {
+    return _props.default;
+  }
+});
+Object.defineProperty(exports, "injectNOS", {
+  enumerable: true,
+  get: function get() {
+    return _inject.injectNOS;
+  }
+});
 
 var _props = _interopRequireDefault(require("./props"));
 
-var _nosStore = require("./nosStore");
+var _inject = require("./inject");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var _default = {
-  Consumer: _nosStore.Consumer,
-  injectNOS: _nosStore.injectNOS,
-  nosProps: _props.default
-};
-exports.default = _default;
