+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nosplatform/api-functions",
-  "version": "0.2.2",
+  "version": "0.2.3",
   "description": "nOS API bindings and types",
   "main": "lib/index.js",
   "scripts": {
@@ -32,14 +32,12 @@
   "dependencies": {
     "prop-types": "15.6.1"
   },
-  "peerDependencies": {
-    "react": ">=16.3.0"
-  },
   "devDependencies": {
     "@babel/cli": "^7.0.0-beta.46",
     "@babel/core": "^7.0.0-beta.46",
     "@babel/preset-env": "^7.0.0-beta.46",
     "@babel/preset-react": "^7.0.0-beta.47",
+    "react": ">=16.3.0",
     "prettier": "^1.12.1"
   },
   "prettier": {
